fix(commands): Accidentally marking commands with suggestions as nullable chore(commands): Inline a lot of DSL functions doc(commands): Document some DSL functions
