+++ idofront-commands/src/main/kotlin/com/mineinabyss/idofront/commands/brigadier/IdoArgument.kt
@@ -4,7 +4,6 @@ import kotlin.reflect.KProperty
 
 class IdoArgument<T>(
     val name: String,
-    val default: (IdoCommandContext.() -> T)? = null,
 ) {
     operator fun getValue(thisRef: Any?, property: KProperty<*>): IdoArgument<T> {
         return this
