+++ idofront-commands/src/main/kotlin/com/mineinabyss/idofront/commands/brigadier/IdoCommand.kt
@@ -6,9 +6,7 @@ import com.mineinabyss.idofront.textcomponents.miniMsg
 import com.mojang.brigadier.arguments.ArgumentType
 import com.mojang.brigadier.builder.ArgumentBuilder
 import com.mojang.brigadier.builder.LiteralArgumentBuilder
-import com.mojang.brigadier.context.CommandContext
 import com.mojang.brigadier.suggestion.SuggestionProvider
-import com.mojang.brigadier.suggestion.SuggestionsBuilder
 import com.mojang.brigadier.tree.LiteralCommandNode
 import io.papermc.paper.command.brigadier.CommandSourceStack
 import io.papermc.paper.command.brigadier.Commands
@@ -19,26 +17,6 @@ import org.bukkit.entity.Player
 import org.bukkit.plugin.Plugin
 import kotlin.reflect.KProperty
 
-data class IdoArgumentBuilder<T>(
-    val type: ArgumentType<out T>,
-    val suggestions: (suspend IdoSuggestionsContext.() -> Unit)? = null,
-//    val default: (IdoCommandContext.() -> T)?,
-)
-
-data class IdoSuggestionsContext(
-    val context: CommandContext<CommandSourceStack>,
-    val suggestions: SuggestionsBuilder,
-) {
-    fun suggestFiltering(name: String) {
-        if (name.startsWith(suggestions.remaining, ignoreCase = true))
-            suggestions.suggest(name)
-    }
-
-    fun suggest(list: List<String>) {
-        list.forEach { suggestFiltering(it) }
-    }
-}
-
 @Suppress("UnstableApiUsage")
 @Annotations
 open class IdoCommand(
@@ -48,30 +26,17 @@ open class IdoCommand(
 ) {
     private val renderSteps = mutableListOf<RenderStep>()
 
-    private fun add(step: RenderStep) {
-        renderSteps += step
-    }
-
-    fun <T> ArgumentType<T>.suggests(suggestions: suspend IdoSuggestionsContext.() -> Unit): IdoArgumentBuilder<T?> {
+    fun <T> ArgumentType<T>.suggests(suggestions: suspend IdoSuggestionsContext.() -> Unit): IdoArgumentBuilder<T> {
         return IdoArgumentBuilder(this, suggestions)
     }
 
-    fun <T> ArgumentType<T>.suggests(provider: SuggestionProvider<CommandSourceStack>): IdoArgumentBuilder<T?> {
+    fun <T> ArgumentType<T>.suggests(provider: SuggestionProvider<CommandSourceStack>): IdoArgumentBuilder<T> {
         return IdoArgumentBuilder(this) { provider.getSuggestions(context, suggestions) }
     }
-//
-//    fun <T> ArgumentType<T>.orElse(default: IdoCommandContext.() -> T): DefaultingArg<T> {
-//        return DefaultingArg(this, default)
-//    }
-//
-//    fun <T> ArgumentType<T>.orError(): DefaultingArg<T> {
-//        return DefaultingArg(this, null)
-//    }
 
     operator fun <T> ArgumentType<T>.provideDelegate(t: T, property: KProperty<*>): IdoArgument<T> {
-        val arg = IdoArgument<T>(property.name)
         add(RenderStep.Builder(Commands.argument(property.name, this)))
-        return arg
+        return IdoArgument(property.name)
     }
 
     operator fun <T> IdoArgumentBuilder<T>.provideDelegate(thisRef: Any?, property: KProperty<*>): IdoArgument<T?> {
@@ -87,15 +52,18 @@ open class IdoCommand(
         return IdoArgument(property.name)
     }
 
-    operator fun String.invoke(init: IdoCommand.() -> Unit) {
+    /** Creates a subcommand using [Commands.literal]. */
+    inline operator fun String.invoke(init: IdoCommand.() -> Unit) {
         add(RenderStep.Command(IdoCommand(Commands.literal(this), this, plugin).apply(init)))
     }
 
-    fun requires(init: CommandSourceStack.() -> Boolean) = edit {
+    /** Specifies a predicate for the command to execute further, may be calculated more than once. */
+    inline fun requires(crossinline init: CommandSourceStack.() -> Boolean) = edit {
         requires { init(it) }
     }
 
-    fun executes(run: IdoCommandContext.() -> Unit) = edit {
+    /** Specifies an end node for the command that runs something, only one executes block can run per command execution. */
+    inline fun executes(crossinline run: IdoCommandContext.() -> Unit) = edit {
         executes { context ->
             try {
                 run(IdoCommandContext(context))
@@ -106,14 +74,21 @@ open class IdoCommand(
         }
     }
 
-    fun playerExecutes(run: IdoPlayerCommandContext.() -> Unit) {
+    /** [executes], ensuring the executor is a player. */
+    inline fun playerExecutes(crossinline run: IdoPlayerCommandContext.() -> Unit) {
         executes {
             if (executor !is Player) commandException("<red>This command can only be run by a player.".miniMsg())
             run(IdoPlayerCommandContext(context))
         }
     }
 
-    fun edit(apply: IdoArgBuilder.() -> ArgumentBuilder<*, *>) {
+    @PublishedApi
+    internal fun add(step: RenderStep) {
+        renderSteps += step
+    }
+
+    /** Directly edit the command in Brigadier. */
+    inline fun edit(crossinline apply: IdoArgBuilder.() -> ArgumentBuilder<*, *>) {
         add(RenderStep.Apply { apply() as IdoArgBuilder })
     }
 
@@ -130,18 +105,3 @@ open class IdoCommand(
         return initial.build()
     }
 }
-
-sealed interface RenderedCommand {
-    fun foldLeft(acc: IdoArgBuilder): IdoArgBuilder
-
-    data class Apply(val apply: IdoArgBuilder.() -> Unit) : RenderedCommand {
-        override fun foldLeft(acc: IdoArgBuilder) = acc.apply(apply)
-    }
-
-    data class ThenFold(val initial: IdoArgBuilder, val list: List<RenderedCommand>) : RenderedCommand {
-        override fun foldLeft(acc: IdoArgBuilder) = acc.apply {
-            then(list.fold(initial) { acc, next -> next.foldLeft(acc) })
-        }
-    }
-}
-// acc.then(optional.etc()).etc()
