+++ idofront-commands/src/main/kotlin/com/mineinabyss/idofront/commands/brigadier/RootIdoCommands.kt
@@ -8,9 +8,11 @@ class RootIdoCommands(
     val commands: Commands,
     val plugin: Plugin,
 ) {
-    private val rootCommands = mutableListOf<IdoRootCommand>()
+    @PublishedApi
+    internal val rootCommands = mutableListOf<IdoRootCommand>()
 
-    operator fun String.invoke(aliases: List<String>, description: String? = null, init: IdoRootCommand.() -> Unit) {
+    /** Creates a new subcommand via a [Commands.literal] argument. */
+    inline operator fun String.invoke(aliases: List<String>, description: String? = null, init: IdoRootCommand.() -> Unit) {
         rootCommands += IdoRootCommand(
             Commands.literal(this),
             this,
@@ -20,13 +22,19 @@ class RootIdoCommands(
         ).apply(init)
     }
 
-    operator fun List<String>.invoke(description: String? = null, init: IdoRootCommand.() -> Unit) =
+    /** Creates a new subcommand with aliases via a [Commands.literal] argument. */
+    inline operator fun List<String>.invoke(description: String? = null, init: IdoRootCommand.() -> Unit) =
         firstOrNull()?.invoke(aliases = drop(1), description = description, init = init)
 
+    /** Builder for commands with aliases. */
     operator fun String.div(other: String) = listOf(this, other)
+
+    /** Builder for commands with aliases. */
     operator fun List<String>.div(other: String) = listOf(this) + other
 
-    fun buildEach() {
+    /** Builds and registers each root level command defined in the DSL. */
+    @PublishedApi
+    internal fun buildEach() {
         rootCommands.forEach { command ->
             commands.register(
                 command.build(),
